moved to useNowPlaying
